moved extract_map to utils
